pyproject.toml
[tool.poetry]
name = "oncall"
version = "0.1.0"
description = ""
authors = ["ahnsv <ahnsv@bc.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.1"
django-ninja = "^0.21.0"
django-auth-oidc = "^0.6.0"

[tool.poetry.dev-dependencies]
ruff = "^0.0.269"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
